requirements/chaumes.txt
-r common.txt
appnope==0.1.0
colorclass==2.2.0
decorator==4.0.11
django-extensions==1.7.9
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pip-upgrader==1.4.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyparsing==2.2.0
simplegeneric==0.8.1
six==1.10.0
terminaltables==3.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
